chats-service/requirements.txt
aiofiles==25.1.0
amqp==5.3.1
annotated-types==0.7.0
anyio==4.11.0
argcomplete==1.10.3
async-timeout==5.0.1
beautifulsoup4==4.8.2
billiard==4.2.2
boto3==1.40.52
botocore==1.40.52
celery==5.5.3
certifi==2025.10.5
chardet==3.0.4
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
compressed-rtf==1.0.7
distro==1.9.0
docx2txt==0.9
ebcdic==1.1.1
exceptiongroup==1.3.0
extract-msg==0.28.7
fastapi==0.119.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
IMAPClient==2.1.0
jiter==0.11.0
jmespath==1.0.1
kombu==5.5.4
lxml==6.0.2
olefile==0.47
openai==2.3.0
packaging==25.0
pdfminer.six==20191110
pillow==11.3.0
prompt_toolkit==3.0.52
psycopg==3.2.10
psycopg-binary==3.2.10
psycopg-pool==3.2.6
pycryptodome==3.23.0
pydantic==2.12.2
pydantic_core==2.41.4
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.1.1
python-multipart==0.0.20
python-pptx==0.6.23
redis==5.2.1
s3transfer==0.14.0
six==1.12.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.8
SpeechRecognition==3.8.1
starlette==0.48.0
textract==1.6.5
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.5.0
uvicorn==0.37.0
vine==5.1.0
wcwidth==0.2.14
xlrd==1.2.0
xlsxwriter==3.2.9
